cdk/requirements/base.txt
aws-cdk-lib==2.149.0,<2.150
boto3==1.34.144,<1.35
constructs==10.3.0,<11